requirements.txt
requests==2.31.0
pandas>=2.3.1
python-dotenv==1.0.0
flask==3.0.0
plotly==5.19.0
numpy>=1.24.0
